requirements.txt
pydantic==2.12.5
email-validator==2.3.0
fastapi==0.118.3
uvicorn==0.38.0
python-dotenv==1.2.1
google-adk==1.19.0
google-genai==1.52.0
motor==3.3.2
pymongo==4.6.1
python-multipart==0.0.9
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0